+++ src/test/java/com/global/api/tests/network/vaps/VapsEncryption3DESTests.java
@@ -18,6 +18,7 @@ import com.global.api.serviceConfigs.NetworkGatewayConfig;
 import com.global.api.services.NetworkService;
 import com.global.api.tests.BatchProvider;
 import com.global.api.tests.StanGenerator;
+import com.global.api.tests.testdata.TestCards;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
@@ -34,6 +35,7 @@ public class VapsEncryption3DESTests {
     private EBTTrackData cashTrack;
     private EBTCardData cashCard;
     private EBTTrackData foodCard;
+    private GiftCard giftCard;
 
     public VapsEncryption3DESTests() throws ApiException {
         Address address = new Address();
@@ -131,6 +133,18 @@ public class VapsEncryption3DESTests {
         foodCard.setPinBlock("62968D2481D231E1A504010024A00014");
         foodCard.setEncryptedPan("4355567063338");
         foodCard.setTrackNumber(TrackNumber.TrackTwo);
+
+        //Gift card
+        config.setNodeIdentification("VLK2");
+        ServicesContainer.configureService(config, "ValueLink");
+
+        // VALUE LINK
+        giftCard = TestCards.ValueLinkSwipe();
+        giftCard.setEncryptionData(EncryptionData.setKtbAndKsn("D87A55F042D1DA9DAD3959DAAE8C3A423E27412D58669AA86993049F07662E478E75B439D9279790",
+                "F0000100095E6701"));
+        giftCard.setEncryptedPan("49FE802FA87C5984BBE68AE7A3277200");
+        giftCard.setTrackNumber(TrackNumber.TrackTwo);
+        giftCard.setEntryMethod(EntryMethod.Swipe);
 
     }
 
@@ -658,5 +672,307 @@ public class VapsEncryption3DESTests {
         // check response
         assertEquals("000", captureResponse.getResponseCode());
     }
+    @Test
+    public void test_force_dataCollect() throws ApiException {
 
+        Transaction response = track.authorize(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute();
+        assertNotNull(response);
+
+        Transaction captureResponse = response.capture()
+                .withCurrency("USD")
+                .withForceToHost(true)
+                .execute();
+        assertNotNull(captureResponse);
+
+        assertNotNull(captureResponse);
+        assertEquals("000", captureResponse.getResponseCode());
+    }
+    @Test
+    public void test_001_resubmitDataCollect_issue_10292() throws ApiException {
+
+        Transaction response = track.authorize(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute();
+        assertNotNull(response);
+
+        Transaction captureResponse = response.capture()
+                .withCurrency("USD")
+                .execute();
+        assertNotNull(captureResponse);
+
+        assertNotNull(captureResponse);
+        assertEquals("000", captureResponse.getResponseCode());
+
+        Transaction resubmit = NetworkService.resubmitDataCollect(captureResponse.getTransactionToken())
+                .withForceToHost(true)
+                .execute();
+        assertNotNull(resubmit);
+        assertEquals("000", resubmit.getResponseCode());
+    }
+    @Test
+    public void test_001_resubmitDataCollectForce_issue_10292() throws ApiException {
+
+        Transaction response = track.authorize(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute();
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+
+        Transaction captureResponse = response.capture()
+                .withCurrency("USD")
+                .execute();
+        assertNotNull(captureResponse);
+
+        assertNotNull(captureResponse);
+        assertEquals("000", captureResponse.getResponseCode());
+
+        Transaction resubmit = NetworkService.resubmitDataCollect(captureResponse.getTransactionToken())
+                .withForceToHost(true)
+                .execute();
+        assertNotNull(resubmit);
+        assertEquals("000", resubmit.getResponseCode());
+    }
+
+
+//-----------------------------------------------GiftCard----------------------------------------
+
+    @Test
+    public void test_001_GiftCard_auth() throws ApiException {
+        Transaction response = giftCard.authorize(new BigDecimal(10),true)
+                .withCurrency("USD")
+                .execute("ValueLink");
+
+        PriorMessageInformation pmi = response.getMessageInformation();
+        assertNotNull(pmi);
+        assertEquals("1100",pmi.getMessageTransactionIndicator());
+        assertEquals("006000",pmi.getProcessingCode());
+        assertEquals("101",pmi.getFunctionCode());
+
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+    }
+
+    @Test
+    public void test_003_giftCard_sale() throws ApiException {
+        Transaction response = giftCard.charge(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+
+        PriorMessageInformation pmi = response.getMessageInformation();
+        assertNotNull(pmi);
+        assertEquals("1200",pmi.getMessageTransactionIndicator());
+        assertEquals("006000",pmi.getProcessingCode());
+        assertEquals("200",pmi.getFunctionCode());
+
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+    }
+
+    @Test
+    public void giftCard_add_value() throws ApiException {
+        Transaction response = giftCard.addValue(new BigDecimal(25.00))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        PriorMessageInformation pmi = response.getMessageInformation();
+        assertNotNull(pmi);
+        assertEquals("1200",pmi.getMessageTransactionIndicator());
+        assertEquals("210060",pmi.getProcessingCode());
+        assertEquals("200",pmi.getFunctionCode());
+
+        assertNotNull(response);
+        assertEquals("000", response.getResponseCode());
+    }
+
+    @Test
+    public void test_GiftCard_preAuthCompletion() throws ApiException {
+        Transaction response = giftCard.authorize(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        PriorMessageInformation pmi = response.getMessageInformation();
+        assertNotNull(pmi);
+        assertEquals("1100",pmi.getMessageTransactionIndicator());
+        assertEquals("006000",pmi.getProcessingCode());
+        assertEquals("100",pmi.getFunctionCode());
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+
+        Transaction captureResponse = response.preAuthCompletion(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(captureResponse);
+        pmi = captureResponse.getMessageInformation();
+        assertNotNull(pmi);
+        assertEquals("1220",pmi.getMessageTransactionIndicator());
+        assertEquals("006000",pmi.getProcessingCode());
+        assertEquals("201",pmi.getFunctionCode());
+        // check response
+        assertEquals("000", captureResponse.getResponseCode());
+    }
+
+    @Test
+    public void giftCard_return() throws ApiException {
+        Transaction response = giftCard.refund(new BigDecimal(35.24))
+                .withCurrency("USD")
+                .withClerkId("41256")
+                .execute("ValueLink");
+        PriorMessageInformation pmi = response.getMessageInformation();
+        assertNotNull(pmi);
+        assertEquals("1200",pmi.getMessageTransactionIndicator());
+        assertEquals("200060",pmi.getProcessingCode());
+        assertEquals("200",pmi.getFunctionCode());
+
+        assertNotNull(response);
+        assertEquals("000", response.getResponseCode());
+    }
+
+    @Test
+    public void test_003_giftCard_sale_void() throws ApiException {
+        Transaction response = giftCard.charge(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+
+        // void the transaction test case #7
+        Transaction voidResponse = response.voidTransaction()
+                .execute("ValueLink");
+        assertNotNull(voidResponse);
+        assertEquals(response.getResponseMessage(), "400", voidResponse.getResponseCode());
+    }
+
+    @Test
+    public void test_004_giftCard_sale_reversal() throws ApiException {
+        NtsData ntsData = new NtsData(FallbackCode.Received_IssuerUnavailable,AuthorizerCode.Terminal_Authorized);
+
+        Transaction response = giftCard.charge(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+
+        response.setNtsData(ntsData);
+
+        // void the transaction test case #8
+        Transaction reverseResponse = response.reverse().execute("ValueLink");
+        assertNotNull(reverseResponse);
+        assertEquals(response.getResponseMessage(), "400", reverseResponse.getResponseCode());
+    }
+    @Test
+    public void test_GiftCard_auth_capture() throws ApiException {
+        Transaction response = giftCard.authorize(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+
+        // test_019
+        Transaction captureResponse = response.capture(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(captureResponse);
+
+        // check response
+        assertEquals("000", captureResponse.getResponseCode());
+    }
+
+    @Test
+    public void test_GiftCard_swipe_voice_capture() throws ApiException {
+        Transaction preResponse = giftCard.authorize(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(preResponse);
+        assertEquals(preResponse.getResponseMessage(), "000", preResponse.getResponseCode());
+
+        Transaction transaction = Transaction.fromNetwork(
+                new BigDecimal(10), preResponse.getAuthorizationCode(),
+                new NtsData(FallbackCode.None, AuthorizerCode.Voice_Authorized),
+                giftCard
+        );
+
+        Transaction response = transaction.capture(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(response);
+
+        // check message data
+        PriorMessageInformation pmi = response.getMessageInformation();
+        assertNotNull(pmi);
+        assertEquals("1220", pmi.getMessageTransactionIndicator());
+        assertEquals("006000", pmi.getProcessingCode());
+        assertEquals("201", pmi.getFunctionCode());
+
+        // check response
+        assertEquals("000", response.getResponseCode());
+    }
+
+    @Test
+    public void test_005_GiftCard_refund() throws ApiException {
+        Transaction response = giftCard.refund(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+    }
+
+    @Test
+    public void test_GiftCard_balance_inquiry() throws ApiException {
+        Transaction response = giftCard.balanceInquiry()
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(response);
+
+        PriorMessageInformation pmi = response.getMessageInformation();
+        assertEquals("1100", pmi.getMessageTransactionIndicator());
+        assertEquals("316000", pmi.getProcessingCode());
+        assertEquals("108", pmi.getFunctionCode());
+
+        assertEquals("000", response.getResponseCode());
+    }
+
+    @Test
+    public void value_link_Replenish_void() throws ApiException {
+
+        Transaction response = giftCard.addValue(new BigDecimal(10.00))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(response);
+        assertEquals("000", response.getResponseCode());
+
+        Transaction voidResponse = response.voidTransaction()
+                .execute("ValueLink");
+        assertNotNull(voidResponse);
+        assertEquals("400", voidResponse.getResponseCode());
+    }
+
+    @Test
+    public void test_005_GiftCard_void_return() throws ApiException {
+        Transaction response = giftCard.refund(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+
+        Transaction voidResponse = response.voidTransaction()
+                .execute("ValueLink");
+        assertNotNull(voidResponse);
+        assertEquals("400", voidResponse.getResponseCode());
+    }
+
+    @Test
+    public void test_007_giftCard_refund_capture() throws ApiException {
+
+        Transaction response = giftCard.refund(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(response);
+        assertEquals(response.getResponseMessage(), "000", response.getResponseCode());
+
+        Transaction capture = response.capture(new BigDecimal(10))
+                .withCurrency("USD")
+                .execute("ValueLink");
+        assertNotNull(capture);
+        assertEquals(capture.getResponseMessage(), "000", capture.getResponseCode());
+    }
 }