+++ src/test/java/com/global/api/tests/testdata/TestCards.java
@@ -571,7 +571,7 @@ public class TestCards {
     }
     public static GiftCard ValueLinkSwipe() {
         GiftCard rvalue = new GiftCard();
-        rvalue.setValue("6010561234567890123=25010004000070779628");
+        rvalue.setValue("6010567093430422=25010004000070772219");
         return rvalue;
     }